Sidecar or Monitor? I chose Sidecar
Sidecar is used in the specification (in one place, but not another). I like the sidecar approach because it is general enough to handle almost everything, and monitoring isn't the only function that might need to be implemented. The downside is that it can be abused, but that won't be the specification's fault.