TrafficSplit spec: Change ordering workflow
Right now the workflow ordering with Kubernetes is as mentioned below:

* Create `foobar-v1` deployment and service
* Create `foobar-v2` deployment and service
* Create `TrafficSplit` object sending all of the traffic to `v1`.

The problem with this approach is that since you have a service `foobar`
which sends traffic to all versions of application. Hence when
`foobar-v2` is created some traffic will be sent to it as well, until we
create that `TrafficSplit` object.

But if we change the workflow ordering to following the traffic is not
sent to the new deployment:

* Create `foobar-v1` deployment and service
* Create `TrafficSplit` object sending all of the traffic to `v1`.
* Create `foobar-v2` deployment and service

Signed-off-by: Suraj Deshmukh <suraj@kinvolk.io>