ref(specs): matches -> routes in HTTPRouteGroup
This PR changes spec.matches to spec.routes in HTTPRouteGroup.
This change makes the spec a little clearer and easier to
understand. This change will also to more easily talk about
routes in the Traffic Split spec as well.

This PR also updates the directory structure of the draft branch
to make it clearer what file is being used to develop a specific
API.

Signed-off-by: Michelle Noorali <minooral@microsoft.com>