chore(*): rm self as maintainer
Thank you so much for the opportunity to be a maintainer on this project. I am unable to continue to fulfill my duties as maintainer due to time constraints and different priorities and am stepping down. I appreciate being able to work with all of you on this project and wish you all the best.

Signed-off-by: Michelle Noorali <michelle.noorali@fermyon.com>